refactor(runtime): split server.go into focused sibling files
Pure file-moves; no behavior changes; no signature changes. Existing tests
on the runtime package stay green and coverage stays at 74.8% (above the
60% target documented in the tier-2 tech-debt plan).

Files:
  server.go            (1592 -> 462 lines: type, NewServer, Shutdown, stderrSyncWriter)
  server_lifecycle.go  (Start, startFileSentry)
  server_reload.go     (Reload, cleanup)
  server_state.go      (currentConfig + 6 sibling accessors, refreshRuntimeState)
  server_emitters.go   (liveReceiptEmitter, liveEnvelopeEmitter, liveFileSentryScanner)
  server_builders.go   (buildToolPolicyCfg, buildMCPInputCfg, buildMCPToolCfg,
                        buildMCPChainMatcher, buildMCPCEE)

Coverage profile on the now-isolated helpers:
- builders: 100% on all 5 (config-derived constructors)
- state accessors: 100% on all 8 (read-side trivial, refreshRuntimeState fully exercised)
- emitters: 75-100% (live* helpers + DLP fallback path)
- reload: 79.2% (Reload) + 92% (cleanup) -- inherited from existing reload tests
- lifecycle: 71.1% (Start) + 62.2% (startFileSentry) -- existing integration coverage

`server.go` is now 462 lines (under the 500 ceiling for the type + NewServer
wiring file). Each sibling file carries only the imports it needs after
goimports prune.
